+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.10.0] - 2019-09-17
+
+### Added
+
+- **appSettings** query.
+- **cards.showMyCards** app setting.
+
 ## [1.8.15] - 2019-09-03 [YANKED]
 
 ### Fixed